U8-Garden Validation.
Looked through Olivers merge review, and implemented the validation for the post code not allowing anything but numbers. Right now the user cannot submit anything but the numbers, but no error message is being thrown. I tried really hard to get the message to popup but i think that the problem right now is that there was no error field made for the location.postcode.

- Will merge right now, but I will probably have to pair-program with Bella to fix it as it should be a 5 min fix.