U6002/tipping-transaction
Pair programming with Angelica.
Added a new column in the transaction "claimed" which is true if the transcation has been claimed by the receiver e.i. you are playing a game. Or false if it hasn't e.i. you are tipping a garden.

Now the receiver can not see the transcation but the sender can.